+++ codex-process-feasibility/src/main/resources/fhir/StructureDefinition/codex-extension-dic.xml
@@ -18,7 +18,7 @@
     <abstract value="false"/>
     <context>
         <type value="element"/>
-        <expression value="MeasureReport.group"/>
+        <expression value="Task.output"/>
     </context>
     <type value="Extension"/>
     <baseDefinition value="http://hl7.org/fhir/StructureDefinition/Extension"/>
